Fold Search engine offline stub into main stub, do not need two - SSDK-617
